+++ apps/mobile/src/components/communities/search-box.tsx
@@ -17,6 +17,9 @@ export function CommunitySearchBox({ onChange, value }: Props) {
 
   return (
     <TextBox
+      autoCapitalize="none"
+      autoComplete="off"
+      autoCorrect={false}
       left={
         <Icon
           color={theme.colors.gray.accent}